Add class to appropriate module
