+++ index.html
@@ -65,7 +65,7 @@
     </div>
   </footer>
 
-  <script src="js/script.js"></script>
+  <script type="module" src="js/script.js"></script>
 </body>
 
 </html>
