+++ modules/time.js
@@ -1,0 +1,5 @@
+import datetime from "luxon/src/datetime";
+
+class Class extends datetime{
+
+}
